Move FooterWellcomeLogo to sit alongside the Footer component
