+++ common/views/components/Footer/FooterWellcomeLogo.js
